+++ code/affluence_counter/detector.py
