+++ code/requieremetns.txt
@@ -1,0 +1,3 @@
+opencv-contrib-python==4.5.1.48
+numpy==1.20.2
+wget==3.2
