--- code/src/demo.py
@@ -1,4 +1,0 @@
-from main import process_video
-
-if __name__ == '__main__':
-    process_video("../../../../videos/2_EnterExitCrossingPaths2front.mpg")
